+++ angular/src/lib/list/examples/list-default.component.ts
@@ -4,7 +4,10 @@ import { Component, OnInit } from '@angular/core';
   selector: 'example-list-default',
   template: `
     <div class="medium-4 columns">
-      <cui-list (select)="onSelect($event)">
+      <cui-list
+        [ngClass]="'my-ng-class'"
+        class="custom-class"
+        (select)="onSelect($event)">
         <div cui-list-item label='List Item A'></div>
         <div cui-list-item label='List Item B'></div>
       </cui-list>